Add global arguments to parser
This major enhancment add the parsing of a global argument dictionary
called __skal__ in the derived class. The options in the dictionary is
then availabel using the normal way: "self.args.<name>".

The syntax of the argument dictionary is:

    class MyApp(SkalApp):

        __skal__ = {
            '-a': {'arg': 'value'},
            ('-l', '--long'): {'arg': 'value'}
        }

Note that the key to the dictionary can be either a sring or a tuple
of two strings representing the short and long verison of the argument.