Skal now properly support module commands
There are two flavours of module commands; as flat commands or with a
hirearchy based on the module name. Both are specified in the SkalApp
constructor, either directly or in a subclass. Application description
and version can now also be specified in the constructor. This is
suggested when using only modules without a subclass of SkalApp.